% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soilcarbon}
\alias{soilcarbon}
\alias{y}
\alias{x}
\alias{ll}
\alias{y.test}
\alias{x.test}
\alias{ll.test}
\title{Soil organic carbon and reflectance spectra data}
\format{
\code{y}: Soil organic carbon
\code{x}: Reflectance spectra
\code{ll}: Locations
\code{y.test}: Prediction data of soil organic carbon
\code{x.test}: Prediction data of reflectance spectra
\code{ll.test}: Prediction data of locations

An object of class \code{matrix} (inherits from \code{array}) with 155 rows and 216 columns.

An object of class \code{data.frame} with 155 rows and 2 columns.

An object of class \code{numeric} of length 67.

An object of class \code{matrix} (inherits from \code{array}) with 67 rows and 216 columns.

An object of class \code{data.frame} with 67 rows and 2 columns.
}
\usage{
y

x

ll

y.test

x.test

ll.test
}
\description{
"soilcarbon" dataset is a simulation data of soil organic carbon and reflectance spectra.
}
\author{
Yongze Song \email{yongze.song@postgrad.curtin.edu.au}
}
\keyword{carbon}
\keyword{dataset;}
\keyword{datasets}
\keyword{organic}
\keyword{soil}
