% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wlr.R
\name{wlr}
\alias{wlr}
\alias{print.wlr}
\alias{plot.wlr}
\alias{predict.wlr}
\title{Wavelet linear regression (WLR)}
\usage{
wlr(y, x, wn, ns = 6)
\method{print}{wlr}(x, ...)
\method{plot}{wlr}(x, ...)
\method{predict}{wlr}(object, newdata, ...)
}
\arguments{
\item{y}{A vector of response variable}

\item{x}{A matrix of independent variables. Row: samples; Column: spectra}

\item{wn}{A vector of wavelength of x}

\item{ns}{Number of scale of multiresolution analysis. Default: ns = 6.}

\item{object}{A list of results}

\item{newdata}{A matrix of new independent variables for prediction}

\item{...}{Ignore}
}
\description{
Function for wavelet linear regression.
}
\examples{
data("soilcarbon")
wn <- seq(350, 2500, 10)
# test
wlr1 <- wlr(y = y, x = x[,1], wn = wn, ns = 6)
# all data
\dontrun{wlr1 <- wlr(y = y, x = x, wn = wn, ns = 6)} # about 10s

}
